requirements.txt
django==1.7
mysql-python
django-suit==0.2.24
django-debug-toolbar==1.3.2
sqlparse==0.1.3
pytils==0.3
django-annoying==0.8.3
django-ckeditor==4.5.1
django-cleanup==0.4.2
django-geoip==0.5.2.1